Supreme zone content
**New Content**

**Supreme Dark beasts**
While grinding slayer tasks you will now have a 1/1000 chance to spawn a Supreme Dark beast.
This beast is only attackable to the player who triggered the spawn.
 The Supreme dark beast will drop the following items on its drop table

*Supreme essence(1-3) - 1/2 chance*
*25-50 Boss Slayer tickets - 1/1 chance*
*100k Upgrade token pack - 1/3 chance*
*$1 Scroll - 1/15 chance*
*$5 Bond - 1/50 chance*
*Deity Attachment - 1/25 chance*
*Scroll of Insurance - 1/25 chance*
*100% DR Scroll (30min) - 1/75 chance*
*2X DMG Scroll (30min) - 1/75 chance*
*Supreme Apendage - 1/250 chance*

**Supreme Essence**
Supreme Essence is dropped from the Supreme Dark beast and 25 of them are required to upgrade the new Supreme unlock scroll.

**Supreme apendage **
Supreme apendage is a rare drop from the Supreme dark beast and is needed for the new Supreme unlock scroll.

**Supreme unlock scroll**
Supreme unlock scroll can be crafted in the max upgrade interface and  will grant permanent access to the command ::supreme and will be consumed on use.

**Supreme Energy**
Supreme Energy is untradeable and can only be obtained by sacrificing items inside the Supreme Lair to the altar.
Supreme energy can be used on weapons to increase your weapons damage 20% of your max hit while in combat. Changing weapons will instantly move the charges to the weapon you're using

**Supreme lair**
After sacrificing items, you are instantly teleblocked, and  the cieling will begin to collapse. Dark wizards will spawn and start to attack you. The only way to exit this area is to use the portal South of the lair. Dying inside of the area will result in all energy in your inventory to be destroyed.